% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComaprarDist.R
\name{distribution_results_distintos}
\alias{distribution_results_distintos}
\title{library(Hitandrun)}
\usage{
distribution_results_distintos(mat1, mat2)
}
\arguments{
\item{two}{dycon matrices}
}
\description{
library(Hitandrun)
}
\examples{

distribution_results_distintos(Matriz1, Matriz2)
distribution_results_iguales(Matriz1, Matriz2)
}
\references{
https://github.com/ComunidadBioInfo/rbioc18/blob/master/materials/collaborative%20projects/projects_descriptions/04_metaboloma.pdf
}
\author{
Felipe de Jesus Muñoz González, Francisco Romero, Mirna Vazquez, Patricia Carvajal
}
