% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampler.R
\name{init.sampler}
\alias{init.sampler}
\title{Transform the structure to polygon constraints}
\usage{
init.sampler(mat, rh, transf = NULL)
}
\arguments{
\item{mat}{Is the matrix that represnets the flux cone}

\item{transf}{transformation to apply to the matrix}
}
\value{
constraint structure
}
\description{
init.sampler gets a matrix and the right-hand side to estabish the constraints of the flux space
}
\examples{
A <- rbind(c(1,0), c(0,1), c(-1,0), c(0,-1))
B <- c(1, 1, 0, 0)
init.sampler(A,B)
     
}
\author{
Muñoz-González, Felipe <fmunoz@lcg.unam.mx>
}
