% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampler.R
\name{run.sampler}
\alias{run.sampler}
\title{Transform the structure to polygon constraints}
\usage{
run.sampler(cone_constr, num.samples, sampler = "hitandrun")
}
\arguments{
\item{mat}{Is the matrix that represnets the flux cone}

\item{transf}{transformation to apply to the matrix}
}
\value{
constraint structure
}
\description{
init.sampler gets a matrix and the right-hand side to estabish the constraints of the flux space
}
\author{
Muñoz-González, Felipe <fmunoz@lcg.unam.mx>
}
